Extracted order-recording methods into OrderManager class
